half cloud function
Did not upload the cloud function and the program works well but cannot find any new data in the database after click adding new data.
Actually I tried to upload (by manually adding wx-server-sdk) the cloud function, but that does not help. And if I reload (instead of adding new changes) the whole function to the cloud, I would come across a lot of errors of using cloud function.┓( ´∀` )┏